Fix: 유성타기 API 수정

+++ backend/src/main/java/starrynight/api/controller/RoomController.java
@@ -37,7 +37,7 @@ public class RoomController {
 
     @ApiOperation(value = "본인을 제외한 랜덤 ID 조회", notes = "랜덤 아이디를 조회한다.")
     @GetMapping("/random/{id}")
-    public ResponseEntity<Integer> randomRoom(@ApiParam(value = "아이디", required = true) @PathVariable Long id) {
+    public ResponseEntity<Integer> randomRoom(@ApiParam(value = "아이디", required = true) @PathVariable int id) {
         Integer randomId = roomService.randomRoom(id);
         return new ResponseEntity<>(randomId, HttpStatus.OK);
     }
